refactor(user): 증명사진 검증 로직을 FormController에서 DraftFormModal로 이동
