+++ apps/user/src/components/form/DraftFormModal/DraftFormModal.tsx
@@ -2,6 +2,8 @@ import { color } from '@maru/design-system';
 import { CheckInput, Column, Confirm, Text } from '@maru/ui';
 import { useState } from 'react';
 import type { ChangeEvent } from 'react';
+import { useFormProfileValueStore } from '@/stores/form/formProfile';
+import { useSetFormStepStore } from '@/stores/form/formStep';
 
 interface DraftFormModalProps {
   isOpen: boolean;
@@ -12,11 +14,28 @@ interface DraftFormModalProps {
 const DraftFormModal = ({ isOpen, onClose, onConfirm }: DraftFormModalProps) => {
   const [inputValue, setInputValue] = useState('');
   const [isInputValid, setIsInputValid] = useState(false);
+  const profileUrl = useFormProfileValueStore();
+  const setFormStep = useSetFormStepStore();
 
   const handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
     const value = e.target.value;
     setInputValue(value);
     setIsInputValid(value === '확인했습니다');
+  };
+
+  const handleConfirm = () => {
+    if (!isInputValid) return;
+    if (!profileUrl?.downloadUrl) {
+      const result = confirm(
+        '증명사진 업로드가 필요합니다. 지원자 정보 페이지로 이동하시겠습니까?'
+      );
+      if (result) {
+        onClose();
+        setFormStep('지원자정보');
+      }
+      return;
+    }
+    onConfirm();
   };
 
   return (
@@ -52,7 +71,7 @@ const DraftFormModal = ({ isOpen, onClose, onConfirm }: DraftFormModalProps) => 
         </Column>
       }
       onClose={onClose}
-      onConfirm={isInputValid ? onConfirm : () => {}}
+      onConfirm={isInputValid ? handleConfirm : () => {}}
       confirmButtonText="제출"
       confirmButtonStyle={{
         backgroundColor: isInputValid ? color.maruDefault : color.gray500,
