+++ apps/user/src/components/form/FormController/FormController.tsx
@@ -12,26 +12,10 @@ interface FormControllerProps {
 }
 
 const FormController = ({ onPrevious, onNext, step }: FormControllerProps) => {
-  const profileUrl = useFormProfileValueStore();
-  const setFormStep = useSetFormStepStore();
-
-  const handleNext = () => {
-    if (step === '자기소개서' && !profileUrl?.downloadUrl) {
-      const result = confirm(
-        '증명사진 업로드가 필요합니다. 지원자 정보로 이동하시겠습니까?'
-      );
-      if (result) {
-        setFormStep('지원자정보');
-      }
-      return;
-    }
-    onNext();
-  };
-
   return (
     <StyledControllerArea>
       {step === '지원자정보' ? (
-        <Button styleType="PRIMARY" size="MEDIUM" width={150} onClick={handleNext}>
+        <Button styleType="PRIMARY" size="MEDIUM" width={150} onClick={onNext}>
           다음
         </Button>
       ) : (
@@ -39,7 +23,7 @@ const FormController = ({ onPrevious, onNext, step }: FormControllerProps) => {
           <Button styleType="SECONDARY" size="MEDIUM" width={150} onClick={onPrevious}>
             이전
           </Button>
-          <Button styleType="PRIMARY" size="MEDIUM" width={150} onClick={handleNext}>
+          <Button styleType="PRIMARY" size="MEDIUM" width={150} onClick={onNext}>
             {step === '자기소개서' ? <p>작성 완료</p> : <p>다음</p>}
           </Button>
         </StyledFormController>
